- Diary Login and Entry-View works
