--- Diary/Diary_Eintraege/admin.json
@@ -1,1 +1,0 @@
-{"name":"admin","rolle":"","password":"password","messageList":[]}
