--- Diary/Diary_Eintraege/alex.json
@@ -1,1 +1,0 @@
-{"name":"alex","rolle":"","password":"password","messageList":[]}
