--- Diary/Diary_Eintraege/patrick.json
@@ -1,1 +1,0 @@
-{"name":"patrick","rolle":"","password":"password","messageList":[]}
