--- Diary/Diary_Eintraege/pillip.json
@@ -1,1 +1,0 @@
-{"name":"pillip","rolle":"","password":"password","messageList":[]}
