--- Diary/Diary_Eintraege/rene.json
@@ -1,1 +1,0 @@
-{"name":"rene","rolle":"Website-Guy","password":"pass","messageList":[]}
